+++ SharePointFramework/PortfolioWebParts/src/components/PortfolioAggregation/ItemModal/types.ts
@@ -1,0 +1,11 @@
+import { IModalProps, IColumn } from '@fluentui/react'
+
+export interface IItemModalProps extends IModalProps {
+  title: string
+  value: any
+  columns?: IColumn[]
+}
+
+export interface IItemModalState {
+  isOpen?: boolean
+}
